kgo.Fetches: add EachPartition callback iterator
This is a handy alternative to the FetchesRecordIter, I personally
prefer this style and @dcrodman indicated this would be useful as well.

The only con is the callback is being passed an 88 byte sized struct,
meaning it falls into DUFFCOPY and is a bit slower. Ideally the heavy
iteration will be in record iteration.