client.EndTransaction: if offsets were added to transaction, ensure we commit
If a client consumes within a transaction but does not produce, when we
end the transaction, before this commit we actually avoided sending an
EndTxn request. This was because no partition was produced to.

As it turns out, because we consumed, we issue AddOffsetsToTxn.
Internally within Kafka, this adds the __consumer_offsets partition
responsible for this group to the transaction itself, and this falls
into AddPartitionsToTxn, which itself is the logic that actually begins
a transaction.

So, if we consume at all, we need to commit so that we properly end the
transaction. If we do not, we leave the transaction open, and
potentially if we are low volume enough, later time out.