consumer group: updates & lock ordering fixes
This is a companion commit to 89146f6, which realistically _needs_ this
commit because:

  - I did not properly split the work and add to the commit properly
  - this commit fixes some lock ordering problems I noticed while
    writing that commit

This should fix the panic in #24 by at least logging on when it would be
detected and continuing, however the bug itself is still a mystery. The
debug logs about what the balance results were should help, though, if
this crops up again.

There are a few lock ordering fixes in here which are now documented
extensively. Notably, PollFetches needs the consumer mu, and there is a
huge reason as to why.

The prerevoke and revoke logic, and how we ensure things are done before
returning sometimes, is all more extensively documented.

Lastly, all instances of assignPartitions is now properly guarded by the
consumer mutex. Prior, some instances were not.