txn: change what we set offsets to on revoke / error
First, we definitely want to grab the committed offsets before we allow
a revoke. Otherwise, we could be revoked, which for the non-cooperative
case, empties out uncommitted.

Second, we do not want to unconditionally set the state back to the
prior committed state, because we could have already rejoined the group
and have received a new committed state. Instead, we set the state to
the current committed state while under the revoke mu. This ensures the
latest committed state while also preventing another revoke / rejoin.