client: fixes for prior commits
- fixes 89146f6, turns out that even an atomic value cannot save the
shutdown. The atomic is still valuable for metadata, so we keep it.
However, we now add a dedicated assign mutex. Now, a heartbeat grabbing
the consumer mu will not be deadlocked on the consumer mu being grabbed
during unset. Instead, unset will hold the assignMu while waiting, and
the consumer mu is held only briefly. We can avoid the consumer mu
entirely in AssignDirect and AssignGroup because they do not poll nor
call assignPartitions

- aborts aborting buffered records on flush, because this is currently
not working until the sequence number thing is addressed. I knew this
was a risk, but I'd rather fix both at once rather than leaving things
broken until the sequence number fix.

- Fixes a logger key/val mismatch that only passed the val.